# apps/api/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Amazon Monitor API",
    "description": "API for Amazon Keyword Competitor Price and Rank Monitor System",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/dashboard/summary": {
      "get": {
        "summary": "Get Dashboard Summary metrics",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "format": "date" },
            "description": "Date in YYYY-MM-DD format"
          }
        ],
        "responses": {
          "200": { "description": "Dashboard summary details" }
        }
      }
    },
    "/keywords": {
      "get": {
        "summary": "List monitored keywords",
        "responses": {
          "200": { "description": "List of keywords" }
        }
      },
      "post": {
        "summary": "Create keyword monitor",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["keyword"],
                "properties": {
                  "keyword": { "type": "string" },
                  "marketplace": { "type": "string", "default": "amazon.com" },
                  "folderName": { "type": "string" },
                  "status": { "type": "string", "enum": ["enabled", "disabled"], "default": "enabled" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Keyword created" }
        }
      }
    },
    "/keywords/{id}": {
      "patch": {
        "summary": "Update keyword monitor",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "keyword": { "type": "string" },
                  "marketplace": { "type": "string" },
                  "folderName": { "type": "string" },
                  "status": { "type": "string", "enum": ["enabled", "disabled"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Keyword updated" }
        }
      },
      "delete": {
        "summary": "Delete keyword monitor",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "Deleted successfully" }
        }
      }
    },
    "/keywords/{id}/detail": {
      "get": {
        "summary": "Get keyword details, snapshots, changes, and alerts",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } },
          { "name": "date", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Detail information of keyword" }
        }
      }
    },
...
    "/collect/run": {
      "post": {
        "summary": "Trigger keyword collection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["date"],
                "properties": {
                  "keywordId": { "type": "integer", "description": "If provided, only collects this keyword; otherwise, collects all." },
                  "date": { "type": "string", "format": "date" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Collection run task logs" }
        }
      }
    },
...
  }
}